+++ packages/address/src/loader.ts
@@ -6,8 +6,9 @@ import {
 } from './types';
 import query from './graphqlQuery';
 
-const GRAPHQL_ENDPOINT = 'https://country-service.shopifycloud.com/graphql';
-enum GRAPHAL_OPERATION_NAMES {
+export const GRAPHQL_ENDPOINT =
+  'https://country-service.shopifycloud.com/graphql';
+export enum GRAPHQL_OPERATION_NAMES {
   countries = 'countries',
   country = 'country',
 }
@@ -17,51 +18,59 @@ const HEADERS = {
   'Access-Control-Allow-Origin': '*',
 };
 
-export async function loadCountries(locale: string): Promise<Country[]> {
+export const loadCountries: (
+  locale: string,
+) => Promise<Country[]> = memoizeAsync(async (locale: string) => {
   const response = await fetch(GRAPHQL_ENDPOINT, {
     method: 'POST',
     headers: HEADERS,
     body: JSON.stringify({
       query,
-      operationName: GRAPHAL_OPERATION_NAMES.countries,
+      operationName: GRAPHQL_OPERATION_NAMES.countries,
       variables: {
         locale: toSupportedLocale(locale),
       },
     }),
   });
+
   const countries:
     | LoadCountriesResponse
     | ResponseError = await response.json();
+
   if ('errors' in countries) {
     throw new CountryLoaderError(countries);
-  } else {
-    return countries.data.countries;
   }
-}
 
-export async function loadCountry(
+  return countries.data.countries;
+});
+
+export const loadCountry: (
   locale: string,
   countryCode: string,
-): Promise<Country> {
-  const response = await fetch(GRAPHQL_ENDPOINT, {
-    method: 'POST',
-    headers: HEADERS,
-    body: JSON.stringify({
-      query,
-      operationName: GRAPHAL_OPERATION_NAMES.country,
-      variables: {
-        countryCode,
-        locale: toSupportedLocale(locale),
-      },
-    }),
-  });
-  const country: LoadCountryResponse = await response.json();
-  if ('errors' in country) {
-    throw new CountryLoaderError(country);
-  } else {
+) => Promise<Country> = memoizeAsync(
+  async (locale: string, countryCode: string): Promise<Country> => {
+    const response = await fetch(GRAPHQL_ENDPOINT, {
+      method: 'POST',
+      headers: HEADERS,
+      body: JSON.stringify({
+        query,
+        operationName: GRAPHQL_OPERATION_NAMES.country,
+        variables: {
+          countryCode,
+          locale: toSupportedLocale(locale),
+        },
+      }),
+    });
+
+    const country: LoadCountryResponse = await response.json();
+
+    if ('errors' in country) {
+      throw new CountryLoaderError(country);
+    }
+
     return country.data.country;
-  }
-}
+  },
+);
 
 class CountryLoaderError extends Error {
   constructor(errors: ResponseError) {
@@ -71,7 +80,7 @@ class CountryLoaderError extends Error {
 }
 
 const DEFAULT_LOCALE = 'EN';
-const SUPPORTED_LOCALES = [
+export const SUPPORTED_LOCALES = [
   'DA',
   'DE',
   'EN',
@@ -84,7 +93,7 @@ const SUPPORTED_LOCALES = [
   'PT_BR',
 ];
 
-function toSupportedLocale(locale: string) {
+export function toSupportedLocale(locale: string) {
   const supportedLocale = locale.replace(/-/, '_').toUpperCase();
 
   if (SUPPORTED_LOCALES.includes(supportedLocale)) {
@@ -92,4 +101,21 @@ function toSupportedLocale(locale: string) {
   } else {
     return DEFAULT_LOCALE;
   }
+}
+
+type AsyncFunc = (...args: any[]) => Promise<any>;
+interface Cache {
+  [key: string]: Promise<any>;
+}
+
+function memoizeAsync(asyncFunction: AsyncFunc) {
+  const cache: Cache = {};
+
+  return (...args: any[]) => {
+    const stringifiedArgs = JSON.stringify(args);
+    if (!cache[stringifiedArgs]) {
+      cache[stringifiedArgs] = asyncFunction.apply(this, args);
+    }
+    return cache[stringifiedArgs] as Promise<any>;
+  };
 }