+++ packages/react-form-state/src/FormState.ts
@@ -5,17 +5,19 @@ import set from 'lodash/set';
 import {memoize, bind} from 'lodash-decorators';
 
 import {mapObject} from './utilities';
-import {FieldDescriptors, FieldState} from './types';
+import {
+  FieldDescriptors,
+  FieldState,
+  ValueMapper,
+  FieldStates,
+  ValidationFunction,
+} from './types';
 import {List, Nested} from './components';
 
 export interface RemoteError {
   field?: string[] | null;
   message: string;
 }
-
-export type FieldStates<Fields> = {
-  [FieldPath in keyof Fields]: FieldState<Fields[FieldPath]>
-};
 
 type MaybeArray<T> = T | T[];
 type MaybePromise<T> = T | Promise<T>;
@@ -26,15 +28,11 @@ interface SubmitHandler<Fields> {
     | MaybePromise<void>;
 }
 
-export type ValidatorDictionary<Fields> = {
-  [FieldPath in keyof Fields]: MaybeArray<
-    ValidationFunction<Fields[FieldPath], Fields>
-  >
-};
+export type Validator<T, F> = MaybeArray<ValidationFunction<T, F>>;
 
-interface ValidationFunction<Value, Fields> {
-  (value: Value, fields: FieldStates<Fields>): any;
-}
+export type ValidatorDictionary<FieldMap> = {
+  [FieldPath in keyof FieldMap]: Validator<FieldMap[FieldPath], FieldMap>
+};
 
 export interface FormData<Fields> {
   fields: FieldDescriptors<Fields>;
@@ -53,6 +51,7 @@ interface Props<Fields> {
   initialValues: Fields;
   validators?: Partial<ValidatorDictionary<Fields>>;
   onSubmit?: SubmitHandler<Fields>;
+  validateOnSubmit?: boolean;
   children(form: FormDetails<Fields>): React.ReactNode;
 }
 
@@ -122,19 +121,29 @@ export default class FormState<
     };
   }
 
+  public validateForm() {
+    return new Promise(resolve => {
+      this.setState(runAllValidators, () => resolve());
+    });
+  }
+
   private get dirty() {
     return this.state.dirtyFields.length > 0;
   }
 
   private get valid() {
-    const {errors, fields} = this.state;
+    const {errors} = this.state;
 
-    const fieldsWithErrors = Object.keys(fields).filter(fieldPath => {
-      const {error} = fields[fieldPath];
-      return error != null;
+    return !this.hasClientErrors && errors.length === 0;
+  }
+
+  private get hasClientErrors() {
+    const {fields} = this.state;
+
+    return Object.keys(fields).some(fieldPath => {
+      const field = fields[fieldPath];
+      return field.error != null;
     });
-
-    return fieldsWithErrors.length === 0 && errors.length === 0;
   }
 
   private get fields() {
@@ -149,7 +158,7 @@ export default class FormState<
 
   @bind()
   private async submit(event?: Event) {
-    const {onSubmit} = this.props;
+    const {onSubmit, validateOnSubmit} = this.props;
     const {formData, mounted} = this;
 
     if (!mounted) {
@@ -165,6 +174,15 @@ export default class FormState<
     }
 
     this.setState({submitting: true});
+
+    if (validateOnSubmit) {
+      await this.validateForm();
+
+      if (this.hasClientErrors) {
+        this.setState({submitting: false});
+        return;
+      }
+    }
 
     const errors = await onSubmit(formData);
 
@@ -200,16 +218,22 @@ export default class FormState<
 
   private updateField<Key extends keyof Fields>(
     fieldPath: Key,
-    value: Fields[Key],
+    value: Fields[Key] | ValueMapper<Fields[Key]>,
   ) {
     this.setState<any>(({fields, dirtyFields}: State<Fields>) => {
       const field = fields[fieldPath];
-      const dirty = !isEqual(value, field.initialValue);
+
+      const newValue =
+        typeof value === 'function'
+          ? (value as ValueMapper<Fields[Key]>)(field.value)
+          : value;
+
+      const dirty = !isEqual(newValue, field.initialValue);
 
       const updatedField = this.getUpdatedField({
         fieldPath,
         field,
-        value,
+        value: newValue,
         dirty,
       });
 
@@ -318,27 +342,9 @@ export default class FormState<
 
     const {validators = {}} = this.props;
     const {fields} = this.state;
-    const validate = validators[fieldPath];
-
-    if (typeof validate === 'function') {
-      // eslint-disable-next-line consistent-return
-      return validate(value, fields);
-    }
-
-    if (!isArray(validate)) {
-      return;
-    }
-
-    const errors = validate
-      .map(validator => validator(value, fields))
-      .filter(input => input != null);
-
-    if (errors.length === 0) {
-      return;
-    }
 
     // eslint-disable-next-line consistent-return
-    return errors;
+    return runValidator(validators[fieldPath], value, fields);
   }
 
   private updateRemoteErrors(errors: RemoteError[]) {
@@ -386,4 +392,56 @@ function createFormState<Fields>(values: Fields): State<Fields> {
 
 function initialValuesFromFields<Fields>(fields: FieldStates<Fields>): Fields {
   return mapObject(fields, ({initialValue}) => initialValue);
+}
+
+function runValidator<T, F>(
+  validate: Validator<T, F> = () => {},
+  value: T,
+  fields: FieldStates<F>,
+) {
+  if (typeof validate === 'function') {
+    // eslint-disable-next-line consistent-return
+    return validate(value, fields);
+  }
+
+  if (!isArray(validate)) {
+    // eslint-disable-next-line consistent-return
+    return;
+  }
+
+  const errors = validate
+    .map(validator => validator(value, fields))
+    .filter(input => input != null);
+
+  if (errors.length === 0) {
+    // eslint-disable-next-line consistent-return
+    return;
+  }
+
+  // eslint-disable-next-line consistent-return
+  return errors;
+}
+
+function runAllValidators<FieldMap>(
+  state: State<FieldMap>,
+  props: Props<FieldMap>,
+) {
+  const {fields} = state;
+  const {validators} = props;
+
+  if (!validators) {
+    return null;
+  }
+
+  const updatedFields = mapObject(fields, (field, path) => {
+    return {
+      ...field,
+      error: runValidator(validators[path], field.value, fields),
+    };
+  });
+
+  return {
+    ...state,
+    fields: updatedFields,
+  } as State<FieldMap>;
 }