+++ packages/react-form-state/src/tests/FormState.test.tsx
@@ -96,9 +96,7 @@ describe('<FormState />', () => {
       const renderPropSpy = jest.fn(() => null);
       const product = faker.commerce.productName();
 
-      const form = mount(
-        <FormState initialValues={{product}}>{renderPropSpy}</FormState>,
-      );
+      mount(<FormState initialValues={{product}}>{renderPropSpy}</FormState>);
 
       const formDetails = lastCallArgs(renderPropSpy);
       formDetails.fields.product.onChange(faker.commerce.productName());
@@ -122,9 +120,7 @@ describe('<FormState />', () => {
       const renderPropSpy = jest.fn(() => null);
       const product = faker.commerce.productName();
 
-      const form = mount(
-        <FormState initialValues={{product}}>{renderPropSpy}</FormState>,
-      );
+      mount(<FormState initialValues={{product}}>{renderPropSpy}</FormState>);
 
       const formDetails = lastCallArgs(renderPropSpy);
       formDetails.fields.product.onChange(faker.commerce.productName());
@@ -168,7 +164,7 @@ describe('<FormState />', () => {
       const product = faker.commerce.productName();
       const color = faker.commerce.color();
 
-      const form = mount(
+      mount(
         <FormState initialValues={{product, color}}>{renderPropSpy}</FormState>,
       );
 
@@ -187,7 +183,7 @@ describe('<FormState />', () => {
       const product = faker.commerce.productName();
       const color = faker.commerce.color();
 
-      const form = mount(
+      mount(
         <FormState initialValues={{product, color}}>{renderPropSpy}</FormState>,
       );
 
@@ -204,9 +200,7 @@ describe('<FormState />', () => {
       const renderPropSpy = jest.fn(() => null);
       const product = faker.commerce.productName();
 
-      const form = mount(
-        <FormState initialValues={{product}}>{renderPropSpy}</FormState>,
-      );
+      mount(<FormState initialValues={{product}}>{renderPropSpy}</FormState>);
 
       const formDetails = lastCallArgs(renderPropSpy);
       const otherProduct = faker.commerce.productName();
@@ -220,9 +214,7 @@ describe('<FormState />', () => {
       const renderPropSpy = jest.fn(() => null);
       const product = faker.commerce.productName();
 
-      const form = mount(
-        <FormState initialValues={{product}}>{renderPropSpy}</FormState>,
-      );
+      mount(<FormState initialValues={{product}}>{renderPropSpy}</FormState>);
 
       const formDetails = lastCallArgs(renderPropSpy);
       formDetails.fields.product.onChange(faker.commerce.productName());
@@ -235,9 +227,7 @@ describe('<FormState />', () => {
       const renderPropSpy = jest.fn(() => null);
       const product = faker.commerce.productName();
 
-      const form = mount(
-        <FormState initialValues={{product}}>{renderPropSpy}</FormState>,
-      );
+      mount(<FormState initialValues={{product}}>{renderPropSpy}</FormState>);
 
       const formDetails = lastCallArgs(renderPropSpy);
       formDetails.fields.product.onChange(faker.commerce.productName());
@@ -250,9 +240,7 @@ describe('<FormState />', () => {
       const renderPropSpy = jest.fn(() => null);
       const product = faker.commerce.productName();
 
-      const form = mount(
-        <FormState initialValues={{product}}>{renderPropSpy}</FormState>,
-      );
+      mount(<FormState initialValues={{product}}>{renderPropSpy}</FormState>);
 
       const formDetails = lastCallArgs(renderPropSpy);
       formDetails.fields.product.onChange(product);
@@ -265,9 +253,7 @@ describe('<FormState />', () => {
       const renderPropSpy = jest.fn(() => null);
       const product = faker.commerce.productName();
 
-      const form = mount(
-        <FormState initialValues={{product}}>{renderPropSpy}</FormState>,
-      );
+      mount(<FormState initialValues={{product}}>{renderPropSpy}</FormState>);
 
       const formDetails = lastCallArgs(renderPropSpy);
       formDetails.fields.product.onChange(faker.commerce.productName());
@@ -281,7 +267,7 @@ describe('<FormState />', () => {
       const product = faker.commerce.productName();
       const description = faker.lorem.words();
 
-      const form = mount(
+      mount(
         <FormState initialValues={{product, description}}>
           {renderPropSpy}
         </FormState>,
@@ -299,7 +285,7 @@ describe('<FormState />', () => {
       const product = faker.commerce.productName();
       const description = faker.lorem.words();
 
-      const form = mount(
+      mount(
         <FormState initialValues={{product, description}}>
           {renderPropSpy}
         </FormState>,
@@ -320,7 +306,7 @@ describe('<FormState />', () => {
       const productValidatorSpy = jest.fn();
       const product = faker.commerce.productName();
 
-      const form = mount(
+      mount(
         <FormState
           initialValues={{
             product,
@@ -371,7 +357,7 @@ describe('<FormState />', () => {
       const renderPropSpy = jest.fn(() => null);
       const validatorSpy = jest.fn(() => faker.lorem.sentence());
 
-      const form = mount(
+      mount(
         <FormState
           initialValues={{product: faker.commerce.productName()}}
           validators={{product: validatorSpy}}
@@ -393,7 +379,7 @@ describe('<FormState />', () => {
       const error = faker.lorem.sentence();
       const validatorSpy = jest.fn(() => error);
 
-      const form = mount(
+      mount(
         <FormState
           initialValues={{product: faker.commerce.productName()}}
           validators={{product: validatorSpy}}
@@ -419,7 +405,7 @@ describe('<FormState />', () => {
       const validatorSpy = jest.fn(() => error);
       const otherValidatorSpy = jest.fn(() => otherError);
 
-      const form = mount(
+      mount(
         <FormState
           initialValues={{product: faker.commerce.productName()}}
           validators={{product: [validatorSpy, otherValidatorSpy]}}
@@ -445,7 +431,7 @@ describe('<FormState />', () => {
       const renderPropSpy = jest.fn(() => null);
       const validatorSpy = jest.fn();
 
-      const form = mount(
+      mount(
         <FormState
           initialValues={{product: faker.commerce.productName()}}
           validators={{product: validatorSpy}}
@@ -466,7 +452,7 @@ describe('<FormState />', () => {
     it('sets valid to false when any field fails validation', () => {
       const renderPropSpy = jest.fn(() => null);
 
-      const form = mount(
+      mount(
         <FormState
           initialValues={{
             product: faker.commerce.productName(),
@@ -496,7 +482,7 @@ describe('<FormState />', () => {
     it('sets valid to true when every field passes validation', () => {
       const renderPropSpy = jest.fn(() => null);
 
-      const form = mount(
+      mount(
         <FormState
           initialValues={{
             product: faker.commerce.productName(),
@@ -523,7 +509,7 @@ describe('<FormState />', () => {
   });
 
   describe('field submit', () => {
-    it('calls onSubmit() with the current formDetails, other than the submit function, when submit() is called', () => {
+    it('calls onSubmit() with the current formDetails, other than the submit function, when submit() is called', async () => {
       const renderPropSpy = jest.fn(() => null);
       const onSubmitSpy = jest.fn();
       const product = faker.commerce.productName();
@@ -538,13 +524,12 @@ describe('<FormState />', () => {
         renderPropSpy,
       );
 
-      submit();
+      await submit();
 
       expect(onSubmitSpy).toHaveBeenLastCalledWith(
         expect.objectContaining(formData),
       );
     });
-
     it('when onSubmit() returns a promise, re-renders with submitting true while waiting for it to resolve/reject', () => {
       const renderPropSpy = jest.fn(() => null);
       const product = faker.commerce.productName();
@@ -687,15 +672,193 @@ describe('<FormState />', () => {
         </FormState>,
       );
 
-      const {submit, submitting, reset, ...formData} = lastCallArgs(
-        renderPropSpy,
-      );
+      const {submit} = lastCallArgs(renderPropSpy);
 
       const submissionPromise = submit();
       form.unmount();
       expect(async () => {
         await submissionPromise();
       }).not.toThrow();
+    });
+  });
+
+  describe('validateOnSubmit', () => {
+    it('calls all validators on submit when validateOnSubmit is true', async () => {
+      const renderPropSpy = jest.fn(() => null);
+      const productValidatorSpy = jest.fn();
+      const skuValidatorSpy = jest.fn();
+
+      mount(
+        <FormState
+          validateOnSubmit
+          initialValues={{
+            product: faker.commerce.productName,
+            sku: faker.commerce.sku,
+          }}
+          validators={{
+            product: productValidatorSpy,
+            sku: skuValidatorSpy,
+          }}
+          onSubmit={noop}
+        >
+          {renderPropSpy}
+        </FormState>,
+      );
+
+      const {submit} = lastCallArgs(renderPropSpy);
+
+      await submit();
+
+      expect(productValidatorSpy).toBeCalled();
+      expect(skuValidatorSpy).toBeCalled();
+    });
+
+    it('does not call onSubmit when a validator fails and validateOnSubmit is true', async () => {
+      const renderPropSpy = jest.fn(() => null);
+      const submitSpy = jest.fn();
+
+      mount(
+        <FormState
+          validateOnSubmit
+          initialValues={{
+            product: faker.commerce.productName,
+          }}
+          validators={{
+            product() {
+              return 'product bad';
+            },
+          }}
+          onSubmit={submitSpy}
+        >
+          {renderPropSpy}
+        </FormState>,
+      );
+
+      const {submit} = lastCallArgs(renderPropSpy);
+
+      await submit();
+
+      expect(submitSpy).not.toBeCalled();
+    });
+
+    it('resets submitting state when validation on submit fails', async () => {
+      const renderPropSpy = jest.fn(() => null);
+
+      mount(
+        <FormState
+          validateOnSubmit
+          initialValues={{
+            product: faker.commerce.productName,
+          }}
+          validators={{
+            product() {
+              return 'product bad';
+            },
+          }}
+          onSubmit={noop}
+        >
+          {renderPropSpy}
+        </FormState>,
+      );
+
+      const {submit} = lastCallArgs(renderPropSpy);
+
+      await submit();
+
+      const {submitting} = lastCallArgs(renderPropSpy);
+
+      expect(submitting).toBe(false);
+    });
+
+    it('does not call any validators on submit when validateOnSubmit is false', async () => {
+      const renderPropSpy = jest.fn(() => null);
+      const productValidatorSpy = jest.fn();
+      const skuValidatorSpy = jest.fn();
+
+      mount(
+        <FormState
+          validateOnSubmit={false}
+          initialValues={{
+            product: faker.commerce.productName,
+            sku: faker.commerce.sku,
+          }}
+          validators={{
+            product: productValidatorSpy,
+            sku: skuValidatorSpy,
+          }}
+          onSubmit={noop}
+        >
+          {renderPropSpy}
+        </FormState>,
+      );
+
+      const {submit} = lastCallArgs(renderPropSpy);
+
+      await submit();
+
+      expect(productValidatorSpy).not.toBeCalled();
+      expect(skuValidatorSpy).not.toBeCalled();
+    });
+  });
+
+  describe('validateForm', () => {
+    it('calls all validators', () => {
+      const productValidatorSpy = jest.fn();
+      const skuValidatorSpy = jest.fn();
+
+      const form = mount(
+        <FormState
+          initialValues={{
+            product: faker.commerce.productName,
+            sku: faker.commerce.sku,
+          }}
+          validators={{
+            product: productValidatorSpy,
+            sku: skuValidatorSpy,
+          }}
+          onSubmit={noop}
+        >
+          {() => <div />}
+        </FormState>,
+      );
+
+      /*
+        unfortunately enzyme doesn't invoke refs so we can't access the instance the
+        way we would in real application code
+      */
+      (form.instance() as FormState<any>).validateForm();
+
+      expect(productValidatorSpy).toBeCalled();
+      expect(skuValidatorSpy).toBeCalled();
+    });
+
+    it('updates fields when a validator fails', async () => {
+      const renderPropSpy = jest.fn(() => null);
+      const error = 'bad';
+      const productValidatorSpy = jest.fn(() => error);
+
+      const form = mount(
+        <FormState
+          initialValues={{
+            product: faker.commerce.productName,
+          }}
+          validators={{
+            product: productValidatorSpy,
+          }}
+          onSubmit={noop}
+        >
+          {renderPropSpy}
+        </FormState>,
+      );
+
+      /*
+        unfortunately enzyme doesn't invoke refs so we can't access the instance the
+        way we would in real application code
+      */
+      await (form.instance() as FormState<any>).validateForm();
+
+      const {fields} = lastCallArgs(renderPropSpy);
+      expect(fields.product.error).toBe(error);
     });
   });
 
@@ -781,3 +944,5 @@ describe('<FormState />', () => {
     });
   });
 });
+
+function noop() {}