+++ packages/react-form-state/src/validators.ts
@@ -1,11 +1,8 @@
-import {toString, isArray} from 'lodash';
+import toString from 'lodash/toString';
+import isArray from 'lodash/isArray';
 import {mapObject} from './utilities';
 
-interface Matcher<Input> {
-  (input: Input): boolean;
-}
-
-interface Matcher<Input, Fields = never> {
+interface Matcher<Input, Fields> {
   (input: Input, fields: Fields): boolean;
 }
 
@@ -36,10 +33,8 @@ export function isEmptyString(input: string) {
   return input.trim().length < 1;
 }
 
-export function not<Input>(matcher: Matcher<Input>): Matcher<Input>;
-
-export function not<Input, Fields>(matcher: Matcher<Input, Fields>) {
-  return (input: Input, fields: Fields) => !matcher(input, fields);
+export function not<A extends Array<any>, R>(fn: (...a: A) => R) {
+  return (...args: A) => !fn(...args);
 }
 
 export function validateNested<Input extends Object, Fields>(
@@ -99,11 +94,11 @@ export function validateList<Input extends Object, Fields>(
 }
 
 export function validate<Input>(
-  matcher: Matcher<Input>,
+  matcher: Matcher<Input, any>,
   errorContent: ErrorContent,
-): (input: Input) => ErrorContent | void;
+): (input: Input) => ErrorContent | undefined | void;
 
-export function validate<Input, Fields = never>(
+export function validate<Input, Fields>(
   matcher: Matcher<Input, Fields>,
   errorContent: ErrorContent,
 ) {
@@ -117,6 +112,32 @@ export function validate<Input, Fields = never>(
     if (isEmpty(input)) {
       return;
     }
+
+    if (matches) {
+      return;
+    }
+
+    if (typeof errorContent === 'function') {
+      // eslint-disable-next-line consistent-return
+      return errorContent(toString(input));
+    }
+
+    // eslint-disable-next-line consistent-return
+    return errorContent;
+  };
+}
+
+export function validateRequired<Input>(
+  matcher: Matcher<Input, any>,
+  errorContent: ErrorContent,
+): (input: Input) => ErrorContent | undefined | void;
+
+export function validateRequired<Input, Fields>(
+  matcher: Matcher<Input, Fields>,
+  errorContent: ErrorContent,
+) {
+  return (input: Input, fields: Fields) => {
+    const matches = matcher(input, fields);
 
     if (matches) {
       return;
@@ -150,35 +171,11 @@ const validators = {
   },
 
   requiredString(errorContent: ErrorContent) {
-    return (input: string) => {
-      if (not(isEmptyString)(input)) {
-        return;
-      }
-
-      if (typeof errorContent === 'function') {
-        // eslint-disable-next-line consistent-return
-        return errorContent(toString(input));
-      }
-
-      // eslint-disable-next-line consistent-return
-      return errorContent;
-    };
+    return validateRequired(not(isEmptyString), errorContent);
   },
 
   required(errorContent: ErrorContent) {
-    return (input: any) => {
-      if (not(isEmpty)(input)) {
-        return;
-      }
-
-      if (typeof errorContent === 'function') {
-        // eslint-disable-next-line consistent-return
-        return errorContent(toString(input));
-      }
-
-      // eslint-disable-next-line consistent-return
-      return errorContent;
-    };
+    return validateRequired(not(isEmpty), errorContent);
   },
 };