+++ plopfile.js
@@ -31,13 +31,18 @@ module.exports = function(plop) {
       },
       {
         type: 'add',
-        path: 'packages/{{kebabCase name}}/tsconfig.json',
-        templateFile: 'templates/tsconfig.hbs.json',
+        path: 'packages/{{kebabCase name}}/tsconfig.build.json',
+        templateFile: 'templates/tsconfig.build.hbs.json',
       },
       {
         type: 'add',
         path: 'packages/{{kebabCase name}}/README.md',
         templateFile: 'templates/README.hbs.md',
+      },
+      {
+        type: 'add',
+        path: 'packages/{{kebabCase name}}/CHANGELOG.md',
+        templateFile: 'templates/CHANGELOG.hbs.md',
       },
       {
         type: 'add',